requirements.txt
Django==2.1.8
pymysql==0.9.3
djangorestframework==3.8.2
redis==3.3.8
Pillow==5.3.0